requirements.txt
pandas
numpy
scikit-learn
statsmodels
matplotlib
python-dotenv
openai